+++ tests/test_cmem_plugin_base.py
@@ -7,7 +7,7 @@ from cmem_plugin_base.dataintegration.entity import (
     EntityPath,
 )
 from cmem_plugin_base.dataintegration.description import Plugin, PluginParameter
-from dataintegration.discovery import discover_plugins, get_packages
+from cmem_plugin_base.dataintegration.discovery import discover_plugins, get_packages
 
 
 @Plugin(
@@ -22,9 +22,9 @@ from dataintegration.discovery import discover_plugins, get_packages
             name="param1",
             label="My First Parameter",
             description="My first description",
-            default_value="default value"
+            default_value="default value",
         )
-    ]
+    ],
 )
 class OutputOnlyPlugin(WorkflowPlugin):
     """Example Plugin - only Output."""