+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](https://semver.org/)
 
+## [0.0.6] 2022-02-28
+
+### Added
+
+- optional plugin identifier
+
+### Changed
+
+- plugin discovery no for multiple base moduls of a prefix
+
 ## [0.0.2] 2022-02-25
 
 ### Added